pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "raiiaf"
version = "1.0.0"
description = "RAIIAF: A binary artifact format for preserving AI image generation context and latent representations"
readme = "README.md"
requires-python = ">=3.9"

authors = [
  { name = "Anuroop V J", email = "anuroopvj@protonmail.com" }
]

license = "MIT"
license-files = ["LICENSE"]

keywords = [
  "ai-images",
  "artifact-format",
  "latent-tensors",
  "image-generation",
  "provenance",
  "metadata",
]

classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Intended Audience :: Science/Research",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3 :: Only",
  "Operating System :: OS Independent",
  "Topic :: Multimedia :: Graphics",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
  "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
  "click",
  "jsonschema",
  "numpy",
  "pillow",
  "pynvml",
  "torch",
  "typing-extensions",
  "zfpy",
  "zstandard",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0",
  "pytest-cov",
  "ruff>=0.1.0",
  "mypy>=1.0",
  "types-Pillow",
]

[project.urls]
Homepage = "https://pypi.org/project/raiiaf/"
Repository = "https://github.com/AnuroopVJ/raiiaf"
Documentation = "https://anuroopvj.github.io/raiiaf/"

[tool.setuptools.packages.find]
where = ["."]
include = ["raiiaf*"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-v --tb=short"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py39"
select = ["E", "W", "F", "I", "B", "C4", "UP", "SIM", "TCH", "PTH"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.9"
ignore_missing_imports = true
strict = false
warn_unused_configs = true

[tool.coverage.run]
branch = true
source = ["raiiaf"]

[tool.coverage.report]
show_missing = true
skip_covered = false
